update toolbox app
